Fix Delay class to accept float for delay_seconds
The PREFECT_TASK_RUN_TAG_CONCURRENCY_SLOT_WAIT_SECONDS setting returns
a float value, not an int. Preserve precision for retry delays.

+++ src/prefect/server/api/concurrency_limits.py
@@ -372,7 +372,7 @@ class Abort(Exception):
 
 
 class Delay(Exception):
-    def __init__(self, delay_seconds: int, reason: str):
+    def __init__(self, delay_seconds: float, reason: str):
         self.delay_seconds = delay_seconds
         self.reason = reason
